refactor : 전략패턴 적용
