+++ src/main/java/step4/MoveRule.java
@@ -1,0 +1,6 @@
+package step4;
+
+@FunctionalInterface
+public interface MoveRule {
+    boolean passCondition();
+}
