+++ src/test/java/step4/CarTest.java
@@ -1,8 +1,6 @@
 package step4;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -26,14 +24,7 @@ public class CarTest {
 
     @Test
     void getRandomNumTest() {
-        Car car = new Car("123");
-        assertThat(car.getRandomNum()).isGreaterThanOrEqualTo(0).isLessThanOrEqualTo(9);
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"4:true", "3:false"}, delimiter = ':')
-    void isSatisfiedConditionForMoveTest(int number, boolean result) {
-        Car car = new Car("123");
-        assertThat(car.isSatisfiedConditionForMove(number)).isEqualTo(result);
+        RandomMoveRule randomMoveRule = new RandomMoveRule();
+        assertThat(randomMoveRule.getRandomNum()).isGreaterThanOrEqualTo(0).isLessThanOrEqualTo(9);
     }
 }
